rename non-react component files with .ts extension,rename non-next.js files to start with capital letter and move Product.ts file to root/entities
