+++ entities/Product.ts
